Reformat all tests

+++ databricks/koalas/tests/test_stats.py
@@ -23,47 +23,48 @@ from databricks.koalas.testing.utils import ReusedSQLTestCase, SQLTestUtils
 
 
 class StatsTest(ReusedSQLTestCase, SQLTestUtils):
-
     def _test_stat_functions(self, pdf, kdf):
-        functions = ['max', 'min', 'mean', 'sum']
+        functions = ["max", "min", "mean", "sum"]
         for funcname in functions:
             self.assert_eq(getattr(kdf.A, funcname)(), getattr(pdf.A, funcname)())
             self.assert_eq(getattr(kdf, funcname)(), getattr(pdf, funcname)())
 
-        functions = ['std', 'var']
+        functions = ["std", "var"]
         for funcname in functions:
             self.assert_eq(getattr(kdf.A, funcname)(), getattr(pdf.A, funcname)(), almost=True)
             self.assert_eq(getattr(kdf, funcname)(), getattr(pdf, funcname)(), almost=True)
 
         # NOTE: To test skew and kurt, just make sure they run.
         #       The numbers are different in spark and pandas.
-        functions = ['skew', 'kurt']
+        functions = ["skew", "kurt"]
         for funcname in functions:
             getattr(kdf.A, funcname)()
             getattr(kdf, funcname)()
 
     def test_stat_functions(self):
-        pdf = pd.DataFrame({'A': [1, 2, 3, 4],
-                            'B': [1.0, 2.1, 3, 4]})
+        pdf = pd.DataFrame({"A": [1, 2, 3, 4], "B": [1.0, 2.1, 3, 4]})
         kdf = koalas.from_pandas(pdf)
         self._test_stat_functions(pdf, kdf)
 
     def test_stat_functions_multiindex_column(self):
-        arrays = [np.array(['A', 'A', 'B', 'B']),
-                  np.array(['one', 'two', 'one', 'two'])]
-        pdf = pd.DataFrame(np.random.randn(3, 4), index=['A', 'B', 'C'], columns=arrays)
+        arrays = [np.array(["A", "A", "B", "B"]), np.array(["one", "two", "one", "two"])]
+        pdf = pd.DataFrame(np.random.randn(3, 4), index=["A", "B", "C"], columns=arrays)
         kdf = koalas.from_pandas(pdf)
         self._test_stat_functions(pdf, kdf)
 
     def test_abs(self):
-        pdf = pd.DataFrame({'A': [1, -2, 3, -4, 5],
-                            'B': [1., -2, 3, -4, 5],
-                            'C': [-6., -7, -8, -9, 10],
-                            'D': ['a', 'b', 'c', 'd', 'e']})
+        pdf = pd.DataFrame(
+            {
+                "A": [1, -2, 3, -4, 5],
+                "B": [1.0, -2, 3, -4, 5],
+                "C": [-6.0, -7, -8, -9, 10],
+                "D": ["a", "b", "c", "d", "e"],
+            }
+        )
         kdf = koalas.from_pandas(pdf)
         self.assert_eq(kdf.A.abs(), pdf.A.abs())
         self.assert_eq(kdf.B.abs(), pdf.B.abs())
-        self.assert_eq(kdf[['B', 'C']].abs(), pdf[['B', 'C']].abs())
+        self.assert_eq(kdf[["B", "C"]].abs(), pdf[["B", "C"]].abs())
         # self.assert_eq(kdf.select('A', 'B').abs(), pdf[['A', 'B']].abs())
 
     def test_axis_on_dataframe(self):
@@ -72,13 +73,17 @@ class StatsTest(ReusedSQLTestCase, SQLTestUtils):
         # Less than 'compute.shortcut_limit' will execute a shortcut
         # by using collected pandas dataframe directly.
         # now we set the 'compute.shortcut_limit' as 1000 explicitly
-        set_option('compute.shortcut_limit', 1000)
+        set_option("compute.shortcut_limit", 1000)
 
         try:
-            pdf = pd.DataFrame({'A': [1, -2, 3, -4, 5] * 300,
-                                'B': [1., -2, 3, -4, 5] * 300,
-                                'C': [-6., -7, -8, -9, 10] * 300,
-                                'D': [True, False, True, False, False] * 300})
+            pdf = pd.DataFrame(
+                {
+                    "A": [1, -2, 3, -4, 5] * 300,
+                    "B": [1.0, -2, 3, -4, 5] * 300,
+                    "C": [-6.0, -7, -8, -9, 10] * 300,
+                    "D": [True, False, True, False, False] * 300,
+                }
+            )
             kdf = koalas.from_pandas(pdf)
             self.assert_eq(kdf.count(axis=1), pdf.count(axis=1))
             self.assert_eq(kdf.var(axis=1), pdf.var(axis=1))
@@ -90,11 +95,11 @@ class StatsTest(ReusedSQLTestCase, SQLTestUtils):
             self.assert_eq(kdf.skew(axis=1), pdf.skew(axis=1))
             self.assert_eq(kdf.mean(axis=1), pdf.mean(axis=1))
         finally:
-            reset_option('compute.shortcut_limit')
+            reset_option("compute.shortcut_limit")
 
     def test_corr(self):
         # Disable arrow execution since corr() is using UDT internally which is not supported.
-        with self.sql_conf({'spark.sql.execution.arrow.enabled': False}):
+        with self.sql_conf({"spark.sql.execution.arrow.enabled": False}):
             # DataFrame
             # we do not handle NaNs for now
             pdf = pd.util.testing.makeMissingDataframe(0.3, 42).fillna(0)
@@ -117,22 +122,25 @@ class StatsTest(ReusedSQLTestCase, SQLTestUtils):
 
     def test_cov_corr_meta(self):
         # Disable arrow execution since corr() is using UDT internally which is not supported.
-        with self.sql_conf({'spark.sql.execution.arrow.enabled': False}):
-            pdf = pd.DataFrame({'a': np.array([1, 2, 3], dtype='i1'),
-                                'b': np.array([1, 2, 3], dtype='i2'),
-                                'c': np.array([1, 2, 3], dtype='i4'),
-                                'd': np.array([1, 2, 3]),
-                                'e': np.array([1.0, 2.0, 3.0], dtype='f4'),
-                                'f': np.array([1.0, 2.0, 3.0]),
-                                'g': np.array([True, False, True]),
-                                'h': np.array(list('abc'))},
-                               index=pd.Index([1, 2, 3], name='myindex'))
+        with self.sql_conf({"spark.sql.execution.arrow.enabled": False}):
+            pdf = pd.DataFrame(
+                {
+                    "a": np.array([1, 2, 3], dtype="i1"),
+                    "b": np.array([1, 2, 3], dtype="i2"),
+                    "c": np.array([1, 2, 3], dtype="i4"),
+                    "d": np.array([1, 2, 3]),
+                    "e": np.array([1.0, 2.0, 3.0], dtype="f4"),
+                    "f": np.array([1.0, 2.0, 3.0]),
+                    "g": np.array([True, False, True]),
+                    "h": np.array(list("abc")),
+                },
+                index=pd.Index([1, 2, 3], name="myindex"),
+            )
             kdf = koalas.from_pandas(pdf)
             self.assert_eq(kdf.corr(), pdf.corr())
 
     def test_stats_on_boolean_dataframe(self):
-        pdf = pd.DataFrame({'A': [True, False, True],
-                            'B': [False, False, True]})
+        pdf = pd.DataFrame({"A": [True, False, True], "B": [False, False, True]})
         kdf = koalas.from_pandas(pdf)
 
         pd.testing.assert_series_equal(kdf.min(), pdf.min())
@@ -158,9 +166,7 @@ class StatsTest(ReusedSQLTestCase, SQLTestUtils):
         self.assertAlmostEqual(ks.std(), ps.std())
 
     def test_some_stats_functions_should_discard_non_numeric_columns_by_default(self):
-        pdf = pd.DataFrame({'i': [0, 1, 2],
-                            'b': [False, False, True],
-                            's': ['x', 'y', 'z']})
+        pdf = pd.DataFrame({"i": [0, 1, 2], "b": [False, False, True], "s": ["x", "y", "z"]})
         kdf = koalas.from_pandas(pdf)
 
         # min and max do not discard non-numeric columns by default
@@ -179,9 +185,7 @@ class StatsTest(ReusedSQLTestCase, SQLTestUtils):
         self.assertEqual(len(kdf.skew()), len(pdf.skew()))
 
     def test_stats_on_non_numeric_columns_should_be_discarded_if_numeric_only_is_true(self):
-        pdf = pd.DataFrame({'i': [0, 1, 2],
-                            'b': [False, False, True],
-                            's': ['x', 'y', 'z']})
+        pdf = pd.DataFrame({"i": [0, 1, 2], "b": [False, False, True], "s": ["x", "y", "z"]})
         kdf = koalas.from_pandas(pdf)
 
         self.assertEqual(len(kdf.sum(numeric_only=True)), len(pdf.sum(numeric_only=True)))
@@ -194,9 +198,7 @@ class StatsTest(ReusedSQLTestCase, SQLTestUtils):
         self.assertEqual(len(kdf.skew(numeric_only=True)), len(pdf.skew(numeric_only=True)))
 
     def test_stats_on_non_numeric_columns_should_not_be_discarded_if_numeric_only_is_false(self):
-        pdf = pd.DataFrame({'i': [0, 1, 2],
-                            'b': [False, False, True],
-                            's': ['x', 'y', 'z']})
+        pdf = pd.DataFrame({"i": [0, 1, 2], "b": [False, False, True], "s": ["x", "y", "z"]})
         kdf = koalas.from_pandas(pdf)
 
         # the lengths are the same, but the results are different.
